# Use a sync.Pool to reduce GC in BufferedSender