This also meant switching from []byte to *Buffer. []byte can't be used
directly, instead you would have to use a *[]byte. This seemed uglier
than just switching to *Buffer which supports truncating.

These changes have shown to greatly reduce the number of objects my
application is generating.