@cactus 2016-01-09
Did a couple of comparative test runs with -bench and -bencmem.

~~~
benchmark                                 old ns/op     new ns/op     delta
BenchmarkBufferedClientInc-2              1601          1593          -0.50%
BenchmarkBufferedClientIncSample-2        747           753           +0.80%
BenchmarkBufferedClientSetInt-2           1540          1588          +3.12%
BenchmarkBufferedClientSetIntSample-2     745           739           -0.81%

benchmark                                 old allocs     new allocs     delta
BenchmarkBufferedClientInc-2              2              1              -50.00%
BenchmarkBufferedClientIncSample-2        1              0              -100.00%
BenchmarkBufferedClientSetInt-2           2              1              -50.00%
BenchmarkBufferedClientSetIntSample-2     1              0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkBufferedClientInc-2              48            16            -66.67%
BenchmarkBufferedClientIncSample-2        29            19            -34.48%
BenchmarkBufferedClientSetInt-2           32            16            -50.00%
BenchmarkBufferedClientSetIntSample-2     29            19            -34.48%
~~~

Seems pretty reasonable. Thanks!